requirements.txt
kafka-python
psycopg2-binary
streamlit
python-dotenv
fixparser